Fixed Sample value bugs, and changed for 64bit double precision type, to 128bit decimal precision type for less lossy calculations

+++ Vapp/Vapp.Media/Audio/Waveformat.cs
@@ -68,7 +68,7 @@ namespace Vapp.Media.Audio
             for (int index = 0; index < max; index++)
             {
                 int count = 0;
-                double total = 0;
+                decimal total = 0;
 
                 foreach (Channel iChannel in this.Channels)
                 {
